eatgo-customer-api/src/test/java/kr/co/bora/eatgo/application/UserServiceTest.java
package kr.co.bora.eatgo.application;

import kr.co.bora.eatgo.domain.UserRepository;
import org.junit.Before;
import org.junit.Test;
import org.junit.rules.ExpectedException;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import static org.junit.Assert.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;

public class UserServiceTest {

    private UserService userService;

    @Mock
    private UserRepository userRepository;

    @Before
    public void setUp() {
        MockitoAnnotations.initMocks(this);

        userService = new UserService(userRepository);
    }


    @Test
    public void registerUser() {
        String email = "tester";
        String name = "Tester";
        String password = "test";

        userService.registerUser(email, name, password);

        verify(userRepository).save(any());
    }

    @Test(ExpectedException=UserExistedException.class)
    public void registerUserWithExistedEmail() {
        String email = "tester";
        String name = "Tester";
        String password = "test";

        userService.registerUser(email, name, password);

        verify(userRepository).save(any());
    }
}